fix(k8s): fix deserialization for patchBody (#3012)
